Refactor issues in roles and permissions

+++ resources/views/dashboard/contact_us/deleted.blade.php
@@ -18,9 +18,11 @@
                 </a>
             </li>
             <li>
-                <a href="{{ route('contact-us.index') }}" class="text-primary transition-colors hover:text-primary-focus dark:text-accent-light dark:hover:text-accent">
-                    Contact Us >
-                </a>
+                @can('contactUs-list')
+                    <a href="{{ route('contact-us.index') }}" class="text-primary transition-colors hover:text-primary-focus dark:text-accent-light dark:hover:text-accent">
+                        Contact Us >
+                    </a>
+                @endcan
             </li>
             <li>
                 <a>
@@ -34,14 +36,14 @@
     <div class="grid grid-cols-1 gap-4 sm:gap-5 lg:gap-6">
         <div id="table-filter">
             <div class="ac js-enabled" id="ac-4">
-                
+
                 <div class="flex items-center justify-between">
                     <h2 class="text-base font-medium tracking-wide text-slate-700 line-clamp-1 dark:text-navy-100">
                         Deleted Contact Us Entries
                     </h2>
-                    
+
                 </div>
-                
+
                 <div class="card mt-3">
                     <div class="is-scrollbar-hidden min-w-full overflow-x-auto">
                         <table class="is-hoverable w-full text-left">
@@ -87,7 +89,7 @@
                                                         Restore
                                                     </button>
                                                 </form>
-                                        
+
                                                 <!-- Permanent Delete Button -->
                                                 <form action="{{ route('contact-us.forceDelete', $contact->id) }}" method="POST">
                                                     @csrf
@@ -100,20 +102,25 @@
                                             </div> --}}
 
 
+                                            @can('contactUs-restore')
+                                                <a href="{{ route('contact-us.restore', $contact->id) }}"
+                                                    class="mx-2 btn bg-slate-150 font-medium text-slate-800 hover:bg-slate-200 focus:bg-slate-203 active:bg-slate-200/80 dark:bg-navy-500 dark:text-navy-50 dark:hover:bg-navy-450 dark:focus:bg-navy-450 dark:active:bg-navy-450/90 btn h-8 w-8 p-0 text-info hover:bg-info/20 focus:bg-error/20 active:bg-info/25">
+                                                    <i class="fa fa-refresh"></i>
+                                                </a>
+                                            @endcan
 
-                                    <a href="{{ route('contact-us.restore', $contact->id) }}"
-                                        class="mx-2 btn bg-slate-150 font-medium text-slate-800 hover:bg-slate-200 focus:bg-slate-203 active:bg-slate-200/80 dark:bg-navy-500 dark:text-navy-50 dark:hover:bg-navy-450 dark:focus:bg-navy-450 dark:active:bg-navy-450/90 btn h-8 w-8 p-0 text-info hover:bg-info/20 focus:bg-error/20 active:bg-info/25">
-                                        <i class="fa fa-refresh"></i>
-                                    </a>
-                                        <form action="{{ route('contact-us.forceDelete', $contact->id) }}" method="POST"
-                                            onsubmit="return confirm('Are you sure you want to delete this feedback?');">
-                                            @csrf
-                                            @method('DELETE')
-                                            <button type="submit"  
-                                            class="mx-2 btn bg-slate-150 font-medium text-slate-800 hover:bg-slate-200 focus:bg-slate-200 active:bg-slate-200/80 dark:bg-navy-500 dark:text-navy-50 dark:hover:bg-navy-450 dark:focus:bg-navy-450 dark:active:bg-navy-450/90 btn h-8 w-8 p-0 text-error hover:bg-error/20 focus:bg-error/20 active:bg-error/25">
-                                                <i class="fa fa-trash-alt"></i></button>
+                                            @can('contactUs-forceDeleted')
 
-                                        </form>     
+                                                    <form action="{{ route('contact-us.forceDelete', $contact->id) }}" method="POST"
+                                                    onsubmit="return confirm('Are you sure you want to delete this feedback?');">
+                                                    @csrf
+                                                    @method('DELETE')
+                                                    <button type="submit"
+                                                    class="mx-2 btn bg-slate-150 font-medium text-slate-800 hover:bg-slate-200 focus:bg-slate-200 active:bg-slate-200/80 dark:bg-navy-500 dark:text-navy-50 dark:hover:bg-navy-450 dark:focus:bg-navy-450 dark:active:bg-navy-450/90 btn h-8 w-8 p-0 text-error hover:bg-error/20 focus:bg-error/20 active:bg-error/25">
+                                                        <i class="fa fa-trash-alt"></i></button>
+
+                                                </form>
+                                            @endcan
 
                                         </td>
                                     </tr>
